Fixed bugs and tests

+++ backend/app/api.py
@@ -44,7 +44,7 @@ def save_board(data: Boards, current_user: dict = Depends(get_current_active_use
     result = db.users.update_one(
         {"email": email, "boards.name": data.name},
         {"$set": {"boards.$": data.model_dump()}}
-    )    
+    )
     
     if result.matched_count == 0:
         db.users.update_one(
@@ -723,8 +723,12 @@ def load_users(current_user: dict = Depends(get_current_active_user)):
     codes = list(db.codes.find(projection={"_id": False}))
     return {"users": users, "codes": codes}
 
+class SaveCircumstance(BaseModel):
+    title: str
+    description: str
+
 @router.put("/save_circumstance/{cid}")
-def save_edited_circumstance(cid: str, data: Circumstance,
+def save_edited_circumstance(cid: str, data: SaveCircumstance,
                              current_user: dict = Depends(get_current_active_user)):
     db.circumstance.update_one(
         {"_id": ObjectId(cid)},
@@ -733,16 +737,19 @@ def save_edited_circumstance(cid: str, data: Circumstance,
             "description": data.description
         }}
     )
+
 @router.post("/save_circumstance")
-def save_new_circumstance(data: Circumstance, current_user: dict = Depends(get_current_active_user)):
-    new_note = db.circumstance.insert_one({"title": data.title, "description": data.description})
+def save_new_circumstance(data: SaveCircumstance, current_user: dict = Depends(get_current_active_user)):
+    email = current_user["email"]
+    new_note = db.circumstance.insert_one({"title": data.title, "description": data.description, "author": email})
     fetch_new_note = db.circumstance.find_one({"_id": new_note.inserted_id})
     fetch_new_note["_id"] = str(fetch_new_note["_id"])
     return fetch_new_note
 
 @router.get("/circumstances")
-def get_circumstances():
-    circumstances = list(db.circumstance.find())
+def get_circumstances(current_user: dict = Depends(get_current_active_user)):
+    email = current_user["email"]
+    circumstances = list(db.circumstance.find({"author": { "$in": ["default", email] }}))
     for c in circumstances:
         c["_id"] = str(c["_id"])
     return circumstances
@@ -750,4 +757,13 @@ def get_circumstances():
 @router.delete("/circumstance/{circumstance_id}")
 def delete_circumstance(circumstance_id: str,
                         current_user: dict = Depends(get_current_active_user)):
-    db.circumstance.delete_one({"_id": ObjectId(circumstance_id)})
+    email = current_user["email"]
+    result = db.circumstance.delete_one({"_id": ObjectId(circumstance_id), "author": email})
+
+    if result.deleted_count == 0:
+        raise HTTPException(
+            status_code=404,
+            detail="Circumstance not found or you do not have permission to delete it"
+        )
+
+    return {"status": "deleted"}
